+++ Source/Core/Core/HW/GCKeyboardEmu.h
@@ -6,7 +6,7 @@
 
 #include <string>
 
-#include "InputCommon/ControllerEmu.h"
+#include "InputCommon/ControllerEmu/ControllerEmu.h"
 
 struct KeyboardStatus;
