+++ Source/Core/Core/HW/GCPadEmu.h
@@ -6,7 +6,7 @@
 
 #include <string>
 
-#include "InputCommon/ControllerEmu.h"
+#include "InputCommon/ControllerEmu/ControllerEmu.h"
 
 class ControlGroup;
