+++ Source/Core/InputCommon/ControllerEmu/ControllerEmu.h
