+++ Source/Core/InputCommon/InputCommon.vcxproj
@@ -35,7 +35,7 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <ItemGroup>
-    <ClCompile Include="ControllerEmu.cpp" />
+    <ClCompile Include="ControllerEmu\ControllerEmu.cpp" />
     <ClCompile Include="ControllerInterface\ControllerInterface.cpp" />
     <ClCompile Include="ControllerInterface\Device.cpp" />
     <ClCompile Include="ControllerInterface\DInput\DInput.cpp" />
@@ -55,7 +55,7 @@
     <ClCompile Include="InputConfig.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="ControllerEmu.h" />
+    <ClInclude Include="ControllerEmu\ControllerEmu.h" />
     <ClInclude Include="ControllerInterface\ControllerInterface.h" />
     <ClInclude Include="ControllerInterface\Device.h" />
     <ClInclude Include="ControllerInterface\DInput\DInput.h" />
